requirements.txt
django>=1.10
django-cas-ng>=3.5.6
django-compressor>=2.1
django-formtools>=2.0
django-libsass>=0.7
django-localflavor-br>=1.0
django-pagseguro2>=1.1.1
factory-boy>=2.8.1
iso8601>=0.1.11
mysqlclient>=1.3.9
pillow>=3.4.2
pytz>=2016.10
qrcode>=5.3
requests>=2.12.1
responses>=0.5.1
